+++ InfiniTAM/ITMLib/Trackers/CPU/ITMExtendedTracker_CPU.h
@@ -32,7 +32,7 @@ namespace ITMLib
 							   float viewFrustum_min,
 							   float viewFrustum_max,
 							   float minColourGradient,
-							   int tukeyCutOff,
+							   float tukeyCutOff,
 							   int framesToSkip,
 							   int framesToWeight,
 							   const ITMLowLevelEngine *lowLevelEngine);
