+++ InfiniTAM/ITMLib/Trackers/CUDA/ITMExtendedTracker_CUDA.h
@@ -39,7 +39,7 @@ namespace ITMLib
 								float viewFrustum_min,
 								float viewFrustum_max,
 								float minColourGradient,
-								int tukeyCutOff,
+								float tukeyCutOff,
 								int framesToSkip,
 								int framesToWeight,
 								const ITMLowLevelEngine *lowLevelEngine);
